% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busco.contig.filter.R
\name{busco.contig.filter}
\alias{busco.contig.filter}
\title{BUSCO contig filter}
\usage{
busco.contig.filter(
  contig.info = NULL,
  compleasm.out = NULL,
  busco.out = NULL,
  out.prefix = NULL
)
}
\arguments{
\item{contig.info}{A two column dataframe containing contig names (first column) and lengths (second column) (REQUIRED)}

\item{compleasm.out}{A dataframe output from compleasm, normally called full_table.tsv (https://github.com/huangnengCSU/compleasm)}

\item{busco.out}{A dataframe output from BUSCO, normally called *_full_table.tsv (https://busco.ezlab.org/)}

\item{out.prefix}{A prefix for the output file. If no prefix is provided, outputs the list of contigs to keep to console (default = NULL; optional)}
}
\description{
Use information of BUSCO gene locations in the genome assembly to create a filter
list of contigs to keep
}
\details{
This function creates a list that can be used by an external program (like seqtk subseq) to subset a
fasta file, keeping the listed sequences. Rules for which contigs are kept are as follows:
1) All contigs that have a BUSCO gene considered to be complete and single (S) copy
2) All contigs that have a BUSCO gene considered to be fragmented (F) or interspaced (I) (only output from compleasm has this)
3) All contigs that do not have ANY BUSCO genes on them
4) The longest contig with duplicated (D) BUSCO genes

The user should supply EITHER the BUSCO output *_full_table.tsv OR a compleasm full_table.tsv. Note that compleasm outputs both formats
and calls the BUSCO format full_table_busco_format.tsv
}
